+++ src/cards/fulcro/democards/card_ui.cljs
@@ -9,6 +9,7 @@
     fulcro.democards.manual-tests-of-dynamic-queries
     fulcro.democards.react-refs
     fulcro.democards.react16-cards
+    fulcro.democards.routing-cards
     fulcro.democards.root-form-refresh
     fulcro.democards.parent-state-refresh
     devcards.core))